claude(skills): fix circular import
