+++ tools/claude/config/skills/scan-git-for-tils/git/commits.py
@@ -6,25 +6,10 @@ import json
 import subprocess
 import sys
 from concurrent.futures import ThreadPoolExecutor, as_completed
-from dataclasses import dataclass
 from datetime import datetime, timedelta
 
 from git.formatting import format_relative_date
-
-
-@dataclass
-class Commit:
-    """A git commit with metadata."""
-
-    hash: str  # Short hash (7 chars)
-    full_hash: str  # Full SHA
-    subject: str  # First line of commit message
-    body: str  # Remaining lines of commit message
-    date: str  # Relative date (e.g., "2 days ago")
-    iso_date: str  # ISO date (YYYY-MM-DD)
-    repo: str  # Repository name (owner/repo)
-    files: list[str]  # Files changed
-    url: str  # GitHub URL
+from git.types import Commit
 
 
 def get_github_username() -> str:
