+++ tools/claude/config/skills/scan-git-for-tils/git/formatting.py
@@ -3,10 +3,8 @@
 from __future__ import annotations
 
 from datetime import datetime
-from typing import TYPE_CHECKING
 
-if TYPE_CHECKING:
-    from git.commits import Commit
+from git.types import Commit
 
 
 def format_relative_date(iso_date: str) -> str:
